+++ ui/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to the **Prowler UI** are documented in this file.
 
-## [1.18.2] (Prowler UNRELEASED)
+## [1.18.2] (Prowler v5.18.2)
 
 ### 🐞 Fixed